requirements.txt

pandas
matplotlib
kafka-python